Actualização do WinForm + alteração das descrições de acordo com os pedidos da Dª Paula
